refactor to separate login for individual pages and header links separately
