+++ app/views/layouts/application.html.erb
@@ -6,19 +6,19 @@
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
     <%= stylesheet_link_tag "tailwind", "inter-font", "data-turbo-track": "reload" %>
-
     <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
     <%= javascript_importmap_tags %>
 
-  <div id="nav" data-controller="format">
-  <% if current_user %>
-    <%= link_to "Edit Profile", edit_user_path(current_user.id), data: { format_target: "link" } %>
-    <%= link_to "Logout", :logout, method: :post, data: { format_target: "link" } %>
-  <% else %>
-    <%= link_to "Register", new_user_path, data: { format_target: "link" } %> |
-    <%= link_to "Login", :login, data: { format_target: "link" } %>
-  <% end %>
-</div>
+    <!-- Navigation Links within the format controller -->
+    <div id="nav" data-controller="format">
+      <% if current_user %>
+        <%= link_to "Edit Profile", edit_user_path(current_user.id), data: { format_target: "link" } %>
+        <%= link_to "Logout", :logout, method: :post, data: { format_target: "link" } %>
+      <% else %>
+        <%= link_to "Register", new_user_path, data: { format_target: "link" } %> |
+        <%= link_to "Login", :login, data: { format_target: "link" } %>
+      <% end %>
+    </div>
   </head>
 
   <body>
